Fix ArticleType[] conflict when extract gainViews() to a unique function
