+++ src/utils/articleType.tsx
@@ -21,31 +21,31 @@ export interface ArticleType {
   readonly _distinctSeqID?: number | undefined;
 }
 
-export interface PublishedDate {
-  seconds: number;
-  nanoseconds?: number;
-}
+// export interface PublishedDate {
+//   seconds: number;
+//   nanoseconds?: number;
+// }
 
-export interface ArticleTypeAlgolia {
-  author: string | null;
-  category: string;
-  briefContent: string | null;
-  country: string;
-  description: string | null;
-  id: string;
-  publishedAt?: number;
-  source: { id: string | null; name: string | null };
-  title: string;
-  url: string;
-  urlToImage: string;
-  articleContent: string;
-  clicks: number;
-  readonly objectID: string;
-  readonly _highlightResult?: {} | undefined;
-  readonly _snippetResult?: {} | undefined;
-  readonly _rankingInfo?: RankingInfo | undefined;
-  readonly _distinctSeqID?: number | undefined;
-}
+// export interface ArticleTypeAlgolia {
+//   author: string | null;
+//   category: string;
+//   briefContent: string | null;
+//   country: string;
+//   description: string | null;
+//   id: string;
+//   publishedAt?: number;
+//   source: { id: string | null; name: string | null };
+//   title: string;
+//   url: string;
+//   urlToImage: string;
+//   articleContent: string;
+//   clicks: number;
+//   readonly objectID: string;
+//   readonly _highlightResult?: {} | undefined;
+//   readonly _snippetResult?: {} | undefined;
+//   readonly _rankingInfo?: RankingInfo | undefined;
+//   readonly _distinctSeqID?: number | undefined;
+// }
 
 export interface ArticleTypeFirestore {
   author: string | null;
